+++ src/main/java/roomescape/controller/ReservationController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.ReservationService;
+import roomescape.service.ReservationService;
 import roomescape.dto.ReservationRequestDto;
 import roomescape.model.Reservation;
