+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -8,16 +8,17 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import roomescape.ReservationService;
+import roomescape.service.ReservationService;
 import roomescape.dto.ReservationTimeRequestDto;
 import roomescape.model.ReservationTime;
+import roomescape.service.ReservationTimeService;
 
 @Controller
 public class ReservationTimeController {
-    private final ReservationService reservationService;
+    private final ReservationTimeService reservationTimeService;
 
-    public ReservationTimeController(ReservationService reservationService) {
-        this.reservationService = reservationService;
+    public ReservationTimeController(ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
     }
 
     @GetMapping("admin/time")
@@ -27,17 +28,17 @@ public class ReservationTimeController {
 
     @PostMapping("/times")
     public ResponseEntity<ReservationTime> addTime(@RequestBody ReservationTimeRequestDto reservationTimeRequestDto) {
-        return ResponseEntity.ok(reservationService.addTime(reservationTimeRequestDto.startAt()));
+        return ResponseEntity.ok(reservationTimeService.addTime(reservationTimeRequestDto.startAt()));
     }
 
     @GetMapping("/times")
     public ResponseEntity<List<ReservationTime>> addTime() {
-        return ResponseEntity.ok(reservationService.getAllTime());
+        return ResponseEntity.ok(reservationTimeService.getAllTime());
     }
 
     @DeleteMapping("/times/{id}")
     public ResponseEntity<Integer> deleteTime(@PathVariable Long id) {
-        return ResponseEntity.ok(reservationService.deleteTime(id));
+        return ResponseEntity.ok(reservationTimeService.deleteTime(id));
     }
 
 }
