+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -1,0 +1,32 @@
+package roomescape.service;
+
+import java.util.List;
+import org.springframework.stereotype.Service;
+import roomescape.model.ReservationTime;
+import roomescape.repository.ReservationTimeRepository;
+
+@Service
+public class ReservationTimeService {
+
+    private final ReservationTimeRepository reservationTimeRepository;
+
+    public ReservationTimeService(ReservationTimeRepository reservationTimeRepository) {
+        this.reservationTimeRepository = reservationTimeRepository;
+    }
+
+    public ReservationTime addTime(String startAt) {
+        return reservationTimeRepository.addTime(startAt);
+    }
+
+    public List<ReservationTime> getAllTime() {
+        return reservationTimeRepository.getAllTime();
+    }
+
+    public Integer deleteTime(Long id) {
+        return reservationTimeRepository.deleteTime(id);
+    }
+
+    public ReservationTime getReservationTimeById(Long id) {
+        return reservationTimeRepository.getReservationTimeById(id);
+    }
+}
